agent/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv
httpx
pydantic
openai
azure-identity
